RAG-Assignment/requirements.txt
# LangChain Core
langchain>=1.2.0
langchain-community>=0.4.1
langchain-text-splitters>=1.1.0

# LLM Provider (choose one or more)
langchain-openai>=1.1.0          # OpenAI
langchain-google-genai>=4.1.3    # Google Gemini

# Vector Store
langchain-chroma>=1.1.0
pysqlite3-binary>=0.5.0

# Document Processing
unstructured>=0.18.0

# Utilities
python-dotenv>=1.2.1